refactor(#1732): 删 Presentation.AGUI 项目,AGUISseWriter 内聚 Hosting(B) (#1750)
* refactor(#1726): committed-state publication contract 下沉 Foundation.Abstractions 斩断 Projection.Core.Abstractions→Foundation.Core 反向依赖 (#1743)

* refactor(#1725): 移动 CommittedStateEventEnvelope 到 Projection.Core.Abstractions 斩断 Channel.Abstractions→Projection.Core 反向依赖 (#1739)

* 斩断 Channel.Abstractions 到 Projection.Core 的反向依赖 #1725

* 修复 #1725 #1739 审查拒绝项

* 斩断 Projection.Core.Abstractions 反向依赖 #1726

* 纯化 committed-state publication 命名空间 #1726 #1743

* 内聚 AGUI SSE writer 到 Hosting #1732
