+++ aevatar.slnx
@@ -97,7 +97,6 @@
     <Project Path="src\Aevatar.Hosting\Aevatar.Hosting.csproj" />
     <Project Path="src\Aevatar.Mainnet.Host.Api\Aevatar.Mainnet.Host.Api.csproj" />
 	    <Project Path="src\Aevatar.AGUI.Contracts\Aevatar.AGUI.Contracts.csproj" />
-	    <Project Path="src\Aevatar.Presentation.AGUI\Aevatar.Presentation.AGUI.csproj" />
 	    <Project Path="src\Aevatar.Scripting.Abstractions\Aevatar.Scripting.Abstractions.csproj" />
 	    <Project Path="src\Aevatar.Scripting.Application\Aevatar.Scripting.Application.csproj" />
 	    <Project Path="src\Aevatar.Scripting.Core\Aevatar.Scripting.Core.csproj" />
