+++ docs/canon/llm-streaming.md
@@ -60,7 +60,7 @@ owner: eanzhao
 7. `src/workflow/Aevatar.Workflow.Presentation.AGUIAdapter/EventEnvelopeToWorkflowRunEventMapper.cs`
 8. `src/Aevatar.CQRS.Projection.Core/Streaming/ProjectionSessionEventHub.cs`
 
-`AGUIEvent` CLR protobuf contracts are generated by `Aevatar.AGUI.Contracts`; `Aevatar.Presentation.AGUI` is the SSE adapter boundary.
+`AGUIEvent` CLR protobuf contracts are generated by `Aevatar.AGUI.Contracts`; `Aevatar.GAgentService.Hosting.Sse` owns AGUI SSE emission for the hosting endpoints that consume those contracts.
 
 ## 4. 整体拓扑图
