--- src/Aevatar.Presentation.AGUI/Aevatar.Presentation.AGUI.csproj
@@ -1,15 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net10.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-    <AssemblyName>Aevatar.Presentation.AGUI</AssemblyName>
-    <RootNamespace>Aevatar.Presentation.AGUI</RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <FrameworkReference Include="Microsoft.AspNetCore.App" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Aevatar.AGUI.Contracts\Aevatar.AGUI.Contracts.csproj" />
-  </ItemGroup>
-</Project>
