+++ test/Aevatar.Workflow.Host.Api.Tests/AGUISseWriterTests.cs
@@ -1,6 +1,6 @@
 using System.Text.Json;
 using Aevatar.AGUI.Contracts;
-using Aevatar.Presentation.AGUI;
+using Aevatar.GAgentService.Hosting.Sse;
 using Aevatar.Workflow.Abstractions;
 using Aevatar.Workflow.Application.Abstractions.Runs;
 using Aevatar.Workflow.Core;
